+++ DTOS/DTOSProductoFrontBack/DTOTalleEnvio.cs
@@ -10,7 +10,7 @@ namespace DTOS.DTOSProductoFrontBack
     {
         public long Id { get; set; } = -1;
         public long IdProducto { get; set; } = -1;
-        public long Cantidad { get; set; } = 0;
+        public int Cantidad { get; set; } = 0;
         public List<DTOColorEnvio> Colores { get; set; } = new List<DTOColorEnvio>();
         public string NombreTalle { get; set; } = "";
     }
